pyproject.toml
[build-system]
requires = ["setuptools>=65.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "marketdata-api"
version = "1.0.1"
description = "Professional Financial Market Data Platform with CLI and API"
authors = [
    {name = "Robin Jonsson", email = "robin@example.com"}
]
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["finance", "market-data", "transparency", "mifid", "esma", "firds", "fitrs", "cli", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Developers",
    "Topic :: Office/Business :: Financial",
    "Topic :: Database",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0.1",
    "pandas>=2.0.0",
    "SQLAlchemy>=2.0.0",
    "Flask>=2.3.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "alembic>=1.12.0",
    "tenacity>=8.2.0",
    "dicttoxml==1.7.16",
    "flask-restx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "tqdm>=4.66.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "setuptools>=65.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-benchmark>=4.0.0",
    "pytest-xdist>=3.3.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0", 
    "pytest-mock>=3.11.0",
    "pytest-benchmark>=4.0.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
]
sqlserver = [
    "pyodbc>=5.2.0",
]

[project.scripts]
marketdata = "marketdata_api.cli:main"
mapi = "marketdata_api.cli:main"

[project.urls]
Homepage = "https://github.com/RobiinJonsson/MarketDataAPI"
Repository = "https://github.com/RobiinJonsson/MarketDataAPI"
Issues = "https://github.com/RobiinJonsson/MarketDataAPI/issues"
Documentation = "https://github.com/RobiinJonsson/MarketDataAPI/docs"

[tool.setuptools.packages.find]
where = ["src"]
include = ["marketdata_api*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json", "*.sql", "*.html", "*.css", "*.js"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "src/marketdata_api/tests",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "esma: marks tests that interact with ESMA services",
    "slow: marks tests that are slow and might need to be skipped in fast test runs", 
    "schema: marks schema mapping related tests",
    "integration: marks integration tests that require full system setup",
    "unit: marks unit tests that test individual components",
    "upgrade: marks tests related to version upgrade functionality",
    "cli: marks command-line interface tests",
]
